requirements.txt
fire
addict
tqdm
requests
flask
numpy
pandas
h5py
scikit-learn
matplotlib
torch
pytorch-lightning
torchvision
psycopg
psycopg_binary
sqlalchemy
kafka-python
avro
pillow
mlflow==3.1.1
albumentationsx
deepdiff
pylint
pytest
black